+++ src/guards/admin.guard.ts
@@ -1,0 +1,14 @@
+import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
+
+@Injectable()
+export class AdminGuard implements CanActivate {
+
+    constructor() {}
+
+    canActivate(context: ExecutionContext): boolean {
+        const req = context.switchToHttp().getRequest();
+        const user = req.body.authenticatedUser;
+
+        return user.isAdmin;
+    }
+}