+++ src/modules/ingredient/ingredient.controller.ts
@@ -1,6 +1,8 @@
-import { Body, Controller, Get, HttpCode, Post, UsePipes, ValidationPipe } from '@nestjs/common';
+import { Body, Controller, Get, HttpCode, Post, UseGuards, UsePipes, ValidationPipe } from '@nestjs/common';
 import { IngredientService } from './ingredient.service';
 import { CreateIngredientDto } from './ingredient.dto';
+import { AdminGuard } from '../../guards/admin.guard';
+import { AuthenticationGuard } from '../../guards/authentication.guard';
 
 @Controller('ingredients')
 export class IngredientController {
@@ -14,6 +16,7 @@ export class IngredientController {
 
     @Post('/create')
     @HttpCode(201)
+    @UseGuards(AuthenticationGuard, AdminGuard)
     @UsePipes(ValidationPipe)
     public async createIngredient(@Body() body: CreateIngredientDto) {
         return await this.ingredientService.create(body);