+++ src/modules/ingredient/ingredient.module.ts
@@ -5,13 +5,16 @@ import { IngredientController } from './ingredient.controller';
 import { IngredientService } from './ingredient.service';
 import { RedisModule } from '../redis/redis.module';
 import { IngredientRepository } from '../../mongodb/repositories/ingredient.repository';
+import { JwtManagerModule } from '../jwt-manager/jwt-manager.module';
+import { JwtManagerService } from '../jwt-manager/jwt-manager.service';
 
 @Module({
     imports: [
         MongooseModule.forFeature([INGREDIENT_MODEL]),
-        RedisModule
+        RedisModule,
+        JwtManagerModule
     ],
     controllers: [IngredientController],
-    providers: [IngredientService, IngredientRepository],
+    providers: [IngredientService, IngredientRepository, JwtManagerService],
 })
 export class IngredientModule {}