+++ src/modules/jwt-manager/jwt-manager.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { UserDataPayload } from '../auth/auth.interface';
+import { UserAccessTokenPayload, UserRefreshTokenPayload } from '../auth/auth.interface';
 import { JwtService } from '@nestjs/jwt';
 
 @Injectable()
@@ -7,13 +7,27 @@ export class JwtManagerService {
 
     constructor(private readonly jwtService: JwtService) {}
 
-    public async encodeUserData(userDataPayload: UserDataPayload): Promise<string> {
-        const { login } = userDataPayload;
-
-        return await this.jwtService.signAsync(login, { secret: process.env.ACCESS_TOKEN_SECRET });
+    public async generateAccessToken(userDataPayload: UserAccessTokenPayload): Promise<string> {
+        return await this.jwtService.signAsync(userDataPayload, { secret: process.env.ACCESS_TOKEN_SECRET });
     }
 
-    public decodeUserData(jwtCookie: string): string {
-        return this.jwtService.decode(jwtCookie) as string;
+    public async generateRefreshToken(userDataPayload: UserRefreshTokenPayload): Promise<string> {
+        return await this.jwtService.signAsync(userDataPayload, { secret: process.env.REFRESH_TOKEN_SECRET });
+    }
+
+    public async verifyAccessToken(accessToken: string): Promise<UserAccessTokenPayload> {
+        return await this.jwtService.verifyAsync(accessToken, { secret: process.env.ACCESS_TOKEN_SECRET });
+    }
+
+    public async verifyRefreshToken(refreshToken: string): Promise<UserRefreshTokenPayload> {
+        return await this.jwtService.verifyAsync(refreshToken, { secret: process.env.REFRESH_TOKEN_SECRET });
+    }
+
+    public async encodeUserData(userDataPayload: UserAccessTokenPayload): Promise<string> {
+        return await this.jwtService.signAsync(userDataPayload, { secret: process.env.ACCESS_TOKEN_SECRET });
+    }
+
+    public decodeUserData(jwtCookie: string): UserAccessTokenPayload {
+        return this.jwtService.decode(jwtCookie) as UserAccessTokenPayload;
     }
 }